NeuroGenesis4.info

A list of info for 4 neuro data sets sampled, containing: Meissner.inVitro.bulk.Hs, LIBD.AZ.inVitro.bulk.Hs, Geschwind.inVivo.sc.Hs, Jabaudon.inVivo.sc.Mm

Usage

data(NeuroGenesis4)

Format

A list containing 4 dataframes

Examples

data(NeuroGenesis4.info)
head(NeuroGenesis4.info$Meissner.inVitro.bulk.Hs)
#>         IDs                     File Cond Rep X      color
#> ESC.1 ESC.1 hESC_rep1.count.fpkm.txt  ESC   1 1      black
#> ESC.2 ESC.2 hESC_rep2.count.fpkm.txt  ESC   2 1      black
#> NE.1   NE.1   NE_rep1.count.fpkm.txt   NE   1 2 lightpink2
#> NE.2   NE.2   NE_rep2.count.fpkm.txt   NE   2 2 lightpink2
#> ERG.1 ERG.1  ERG_rep1.count.fpkm.txt  ERG   1 3 indianred1
#> ERG.2 ERG.2  ERG_rep2.count.fpkm.txt  ERG   2 3 indianred1
head(NeuroGenesis4.info$LIBD.AZ.inVitro.bulk.Hs)
#>     EXP_STATUS RNAseq_STATUS Order  RNA_NO     SAMPLE_NAME EXPERIMENT
#> 126   Complete      Complete   126 R16-191    165-B-2 D2 1 TIMECOURSE
#> 164   Complete      Complete   164 R16-287  165-B-2 D9 2-1 TIMECOURSE
#> 127   Complete      Complete   127 R16-197    165-B-3 D2 1 TIMECOURSE
#> 469   Complete      Complete    NA R16-850 165-B-3_D2_REP1 TIMECOURSE
#> 470   Complete      Complete    NA R16-851 165-B-3_D6_REP1 TIMECOURSE
#> 165   Complete      Complete   165 R16-288  165-B-3 D9 2-1 TIMECOURSE
#>     RNA_CONC_ng/ul Library LibConstruct   DateToAJ SPECIES      JC_Comment
#> 126             NA  Batch4   2016-03-31 2016-04-13   HUMAN ASSAY_1.1_COND2
#> 164             NA  Batch4   2016-03-31 2016-04-13   HUMAN ASSAY_1.1_COND2
#> 127             NA  Batch4   2016-03-31 2016-04-13   HUMAN ASSAY_1.1_COND2
#> 469          423.4 BATCH10   2016-09-08 2016-09-28   HUMAN ASSAY_1.1_COND2
#> 470          251.5 BATCH10   2016-09-08 2016-09-28   HUMAN ASSAY_1.1_COND2
#> 165             NA  Batch4   2016-03-31 2016-04-13   HUMAN ASSAY_1.1_COND2
#>      Flowcell    Index Lane  NumReads            SeqNotes NumMapped   mapRate
#> 126 H5TKNBBXX ATTCAGAA   L8  70256642                <NA>  64493777 0.9179741
#> 164 H5TKNBBXX CGGCTATG   L2  47653100                <NA>  38805346 0.8143299
#> 127 H5TKNBBXX AGCGATAG   L4  84345206 10.64% undetermined  79142539 0.9383170
#> 469 H7V3TBBXX GAGATTCC   L6 143106660                <NA> 133682681 0.9341472
#> 470 H7V3TBBXX GAATTCGT   L6 174230786                <NA> 161315252 0.9258711
#> 165 H5TKNBBXX TCCGGAGA   L1  76574342                <NA>  66510301 0.8685716
#>     NumProperMap properRate NumUnmapped unmapRate riboMapped    riboRate
#> 126     44317316  0.6871565    10866616 0.1546703    1848394 0.028660036
#> 164     27965698  0.7206661    12397154 0.2601542    2141647 0.055189483
#> 127     54477558  0.6883474    11344493 0.1345007     124652 0.001575032
#> 469     98673366  0.7381163    17993507 0.1257349    2816740 0.021070343
#> 470    122875468  0.7617102    23249854 0.1334429    1977016 0.012255605
#> 165     44773122  0.6731758    15654357 0.2044335    3235748 0.048650329
#>     mitoMapped    mitoRate alignNote         Class Donor    LINE  DX RPS
#> 126     951371 0.014924190      <NA> Naked genomes   165 165-B-2 CNT  lo
#> 164     523056 0.013734300      <NA> Naked genomes   165 165-B-2 CNT  lo
#> 127     434379 0.005578326      <NA> Naked genomes   165 165-B-3 CNT  lo
#> 469     935448 0.007175070      <NA> Naked genomes   165 165-B-3 CNT  lo
#> 470     873678 0.005559844      <NA> Naked genomes   165 165-B-3 CNT  lo
#> 165     796443 0.012205907      <NA> Naked genomes   165 165-B-3 CNT  lo
#>     Manipulation shRNA MOI PURO   SAMPLE_NAME.1 shRNA_Target_Seq Note DAY
#> 126           NO    NA  NA   NA    165-B-2 D2 1               NA <NA>   2
#> 164           NO    NA  NA   NA  165-B-2 D9 2-1               NA <NA>   9
#> 127           NO    NA  NA   NA    165-B-3 D2 1               NA <NA>   2
#> 469           NO    NA  NA   NA 165-B-3_D2_REP1               NA <NA>   2
#> 470           NO    NA  NA   NA 165-B-3_D6_REP1               NA <NA>   6
#> 165           NO    NA  NA   NA  165-B-3 D9 2-1               NA <NA>   9
#>     BIO_REP Fluidigm     CONDITION          SampleID
#> 126       1       NA ACC_DORSAL(2) R16-191_H5TKNBBXX
#> 164       1       NA ACC_DORSAL(2) R16-287_H5TKNBBXX
#> 127       1       NA ACC_DORSAL(2) R16-197_H5TKNBBXX
#> 469       1       NA ACC_DORSAL(2) R16-850_H7V3TBBXX
#> 470       1       NA ACC_DORSAL(2) R16-851_H7V3TBBXX
#> 165       1       NA ACC_DORSAL(2) R16-288_H5TKNBBXX
#>                                                                                    bamFile
#> 126 /dcl01/lieber/ajaffe/Astrazeneca/TopHat/Sample_R16-191_H5TKNBBXX_out/accepted_hits.bam
#> 164 /dcl01/lieber/ajaffe/Astrazeneca/TopHat/Sample_R16-287_H5TKNBBXX_out/accepted_hits.bam
#> 127 /dcl01/lieber/ajaffe/Astrazeneca/TopHat/Sample_R16-197_H5TKNBBXX_out/accepted_hits.bam
#> 469 /dcl01/lieber/ajaffe/Astrazeneca/TopHat/Sample_R16-850_H7V3TBBXX_out/accepted_hits.bam
#> 470 /dcl01/lieber/ajaffe/Astrazeneca/TopHat/Sample_R16-851_H7V3TBBXX_out/accepted_hits.bam
#> 165 /dcl01/lieber/ajaffe/Astrazeneca/TopHat/Sample_R16-288_H5TKNBBXX_out/accepted_hits.bam
#>     leftInput leftMapped leftMapRate rightInput rightMapped rightMapRate
#> 126  35128321   30441590       0.867   35128321    28948436        0.824
#> 164  23826550   17892853       0.751   23826550    17363093        0.729
#> 127  42172603   37200746       0.882   42172603    35799967        0.849
#> 469  71553330   63377397       0.886   71553330    61735756        0.863
#> 470  87115393   76260852       0.875   87115393    74720080        0.858
#> 165  38287171   31423079       0.821   38287171    29496906        0.770
#>     overallMapRate concordMapRate totalMapped totalAssignedGene lineRep
#> 126          0.845          0.782    62795541         0.6497077       2
#> 164          0.740          0.669    37560864         0.6179368       2
#> 127          0.865          0.806    77434676         0.7626106       3
#> 469          0.874          0.825   129439310         0.7630018       3
#> 470          0.867          0.824   156267057         0.7672984       3
#> 165          0.796          0.715    64454176         0.6493536       3
#>     colDonor PROG.21-B-8 PROG.165-B-6X PROG.66-A-3 PROG.90-A-5 PROG.21-B-9
#> 126    black       FALSE         FALSE       FALSE       FALSE       FALSE
#> 164    black       FALSE         FALSE       FALSE       FALSE       FALSE
#> 127    black       FALSE         FALSE       FALSE       FALSE       FALSE
#> 469    black       FALSE         FALSE       FALSE       FALSE       FALSE
#> 470    black       FALSE         FALSE       FALSE       FALSE       FALSE
#> 165    black       FALSE         FALSE       FALSE       FALSE       FALSE
#>     PROG.165-B-8X PROG.66-A-9 PROG.90-A-10 PROG.165-B-2 PROG.165-B-3 PROG.3-A-7
#> 126         FALSE       FALSE        FALSE         TRUE        FALSE      FALSE
#> 164         FALSE       FALSE        FALSE         TRUE        FALSE      FALSE
#> 127         FALSE       FALSE        FALSE        FALSE         TRUE      FALSE
#> 469         FALSE       FALSE        FALSE        FALSE         TRUE      FALSE
#> 470         FALSE       FALSE        FALSE        FALSE         TRUE      FALSE
#> 165         FALSE       FALSE        FALSE        FALSE         TRUE      FALSE
#>     PROG.3-A-8 PROG.165-B-3X PROG.21-B-3  DxRps pchBYdiagRPS     DAYj DAYx
#> 126      FALSE         FALSE       FALSE CNT.lo           19 1.431041    4
#> 164      FALSE         FALSE       FALSE CNT.lo           19 8.883865    7
#> 127      FALSE         FALSE       FALSE CNT.lo           19 1.092173    4
#> 469      FALSE         FALSE       FALSE CNT.lo           19 2.241799    4
#> 470      FALSE         FALSE       FALSE CNT.lo           19 6.932188    6
#> 165      FALSE         FALSE       FALSE CNT.lo           19 9.986775    7
#>     labelsX colorBYlabelsX
#> 126       2            red
#> 164       9            red
#> 127       2            red
#> 469       2            red
#> 470       6            red
#> 165       9            red
head(NeuroGenesis4.info$Geschwind.inVivo.sc.Hs)
#>             Cell Cluster Subcluster Donor Layer Gestation_week Index Library
#> 169 TGCTAATACTGA     vRG      vRG_0   368    CP             17  N701   Plath
#> 624 TTCACGATTTTT     vRG      vRG_2   368    GZ             17  N702   Plath
#> 659 CTGTCAGAATAA     vRG      vRG_2   368    GZ             17  N702   Plath
#> 681 CATAATATGTCA     vRG      vRG_0   368    GZ             17  N702   Plath
#> 715 TGCCAATCCGTT     vRG      vRG_0   368    GZ             17  N702   Plath
#> 723 TCGACTATTTTG     vRG      vRG_2   368    GZ             17  N702   Plath
#>     Number_genes_detected Number_UMI Percentage_mitochondrial S_phase_score
#> 169                  1366       2736                     4.20         -0.11
#> 624                  3012       7318                     1.87          0.83
#> 659                  1597       3451                     2.17         -0.13
#> 681                  1551       2941                     1.43          0.38
#> 715                  1569       2769                     1.81         -0.04
#> 723                  1610       2994                     3.57          0.11
#>     G2M_phase_score Phase            X    tSNE_1   tSNE_2
#> 169          -0.073    G1 TGCTAATACTGA -3.446397 17.68239
#> 624           0.350     S TTCACGATTTTT -8.190806 22.31286
#> 659          -0.110    G1 CTGTCAGAATAA -3.301333 18.49425
#> 681           0.064     S CATAATATGTCA -8.502828 16.35404
#> 715          -0.170    G1 TGCCAATCCGTT -3.293814 19.02176
#> 723           0.064     S TCGACTATTTTG -7.173418 21.53491
head(NeuroGenesis4.info$Jabaudon.inVivo.sc.Mm)
#>                         title geo_accession                status
#> GSM3351837  E14.24H_LTHT7_1_1    GSM3351837 Public on May 10 2019
#> GSM3351838  E14.24H_LTHT7_1_2    GSM3351838 Public on May 10 2019
#> GSM3351839  E14.24H_LTHT7_1_5    GSM3351839 Public on May 10 2019
#> GSM3351840  E14.24H_LTHT7_1_6    GSM3351840 Public on May 10 2019
#> GSM3351841  E14.24H_LTHT7_1_7    GSM3351841 Public on May 10 2019
#> GSM3351842 E14.24H_LTHT7_1_12    GSM3351842 Public on May 10 2019
#>            submission_date last_update_date type channel_count source_name_ch1
#> GSM3351837     Aug 23 2018      May 10 2019  SRA             1  cortical cells
#> GSM3351838     Aug 23 2018      May 10 2019  SRA             1  cortical cells
#> GSM3351839     Aug 23 2018      May 10 2019  SRA             1  cortical cells
#> GSM3351840     Aug 23 2018      May 10 2019  SRA             1  cortical cells
#> GSM3351841     Aug 23 2018      May 10 2019  SRA             1  cortical cells
#> GSM3351842     Aug 23 2018      May 10 2019  SRA             1  cortical cells
#>            organism_ch1 characteristics_ch1 characteristics_ch1.1
#> GSM3351837 Mus musculus       tsne1: -21.73          tsne2: 10.41
#> GSM3351838 Mus musculus       tsne1: -12.66          tsne2: -0.39
#> GSM3351839 Mus musculus         tsne1: 4.78           tsne2: 2.06
#> GSM3351840 Mus musculus       tsne1: -14.95           tsne2: 7.12
#> GSM3351841 Mus musculus       tsne1: -15.24           tsne2: 7.04
#> GSM3351842 Mus musculus       tsne1: -19.39          tsne2: 11.40
#>            characteristics_ch1.2 characteristics_ch1.3 characteristics_ch1.4
#> GSM3351837      num_read: 608677          pct_mt: 4.3%     cluster: N1d.late
#> GSM3351838       num_read: 68123          pct_mt: 4.8%     cluster: N1d.late
#> GSM3351839      num_read: 608394          pct_mt: 3.5%     cluster: N1d.late
#> GSM3351840      num_read: 557976          pct_mt: 3.4%     cluster: N1d.late
#> GSM3351841      num_read: 121333          pct_mt: 8.7%     cluster: N1d.late
#> GSM3351842      num_read: 621786          pct_mt: 2.9%     cluster: N1d.late
#>                characteristics_ch1.5 characteristics_ch1.6
#> GSM3351837 pup.age.at.injection: E14         cell.age: 24H
#> GSM3351838 pup.age.at.injection: E14         cell.age: 24H
#> GSM3351839 pup.age.at.injection: E14         cell.age: 24H
#> GSM3351840 pup.age.at.injection: E14         cell.age: 24H
#> GSM3351841 pup.age.at.injection: E14         cell.age: 24H
#> GSM3351842 pup.age.at.injection: E14         cell.age: 24H
#>                    characteristics_ch1.7      characteristics_ch1.8
#> GSM3351837 pseudo.differentiation: 0.409 pseudo.birthdate24h: 0.417
#> GSM3351838 pseudo.differentiation: 0.381 pseudo.birthdate24h: 0.462
#> GSM3351839 pseudo.differentiation: 0.327 pseudo.birthdate24h: 0.459
#> GSM3351840 pseudo.differentiation: 0.378 pseudo.birthdate24h: 0.505
#> GSM3351841 pseudo.differentiation: 0.372 pseudo.birthdate24h: 0.500
#> GSM3351842 pseudo.differentiation: 0.421 pseudo.birthdate24h: 0.431
#>                   characteristics_ch1.9             characteristics_ch1.10
#> GSM3351837 pseudo.birthdate.pred: 0.333 pseudo.differentiation.pred: 0.592
#> GSM3351838 pseudo.birthdate.pred: 0.152 pseudo.differentiation.pred: 0.130
#> GSM3351839 pseudo.birthdate.pred: 0.890 pseudo.differentiation.pred: 0.134
#> GSM3351840 pseudo.birthdate.pred: 0.439 pseudo.differentiation.pred: 0.119
#> GSM3351841 pseudo.birthdate.pred: 0.049 pseudo.differentiation.pred: 0.056
#> GSM3351842 pseudo.birthdate.pred: 0.578 pseudo.differentiation.pred: 0.172
#>                      characteristics_ch1.11
#> GSM3351837 pseudo.birthdate.adj.pred: 3.411
#> GSM3351838 pseudo.birthdate.adj.pred: 3.434
#> GSM3351839 pseudo.birthdate.adj.pred: 3.963
#> GSM3351840 pseudo.birthdate.adj.pred: 3.316
#> GSM3351841 pseudo.birthdate.adj.pred: 2.687
#> GSM3351842 pseudo.birthdate.adj.pred: 3.553
#>                            characteristics_ch1.12 molecule_ch1
#> GSM3351837 pseudo.differentiation.adj.pred: 2.219    polyA RNA
#> GSM3351838 pseudo.differentiation.adj.pred: 1.982    polyA RNA
#> GSM3351839 pseudo.differentiation.adj.pred: 1.982    polyA RNA
#> GSM3351840 pseudo.differentiation.adj.pred: 1.943    polyA RNA
#> GSM3351841 pseudo.differentiation.adj.pred: 1.919    polyA RNA
#> GSM3351842 pseudo.differentiation.adj.pred: 1.908    polyA RNA
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        extract_protocol_ch1
#> GSM3351837 The putative primary somatosensory cortex S1 was microdissected and incubated in 0.05% trypsin at 37°C for 5 minutes. Following tissue digestion, fetal bovine serum was added to the mix and cells were manually dissociated via up-and-down pipetting. Cells were centrifuged 5min at 300g and the pellet was suspended in 1ml of HBSS then passed on a 70µm cell stainer. FT+ cells, gated on the top 5% brightest cells, were finally FAC-sorted on a MoFloAstrios (Beckman). FAC-sorted FT+ cells (18µl) were mixed with the C1 Suspension Reagent (2µl; Fluidigm) yielding a total of 20µl of cell suspension mix with ~500 cells/μl. The cell suspension mix was loaded on a C1 Single-Cell AutoPrep integrated fluidic circuit (IFC) designed for 10- to 17-µm cells (HT-800, Fluidigm #100-57-80). cDNA synthesis and preamplification was processed following the manufacturer’s instructions (C1 system, Fluidigm). Single cell RNA-sequencing libraries of the cDNA were prepared using Nextera XT DNA library prep kit (Illumina). Libraries were multi-plexed and sequenced according to the manufacturer’s recommendations with paired-end reads using HiSeq2500 platform (Illumina) with an expected depth of 1M reads per single cell, and a final mapping read length of 70bp.
#> GSM3351838 The putative primary somatosensory cortex S1 was microdissected and incubated in 0.05% trypsin at 37°C for 5 minutes. Following tissue digestion, fetal bovine serum was added to the mix and cells were manually dissociated via up-and-down pipetting. Cells were centrifuged 5min at 300g and the pellet was suspended in 1ml of HBSS then passed on a 70µm cell stainer. FT+ cells, gated on the top 5% brightest cells, were finally FAC-sorted on a MoFloAstrios (Beckman). FAC-sorted FT+ cells (18µl) were mixed with the C1 Suspension Reagent (2µl; Fluidigm) yielding a total of 20µl of cell suspension mix with ~500 cells/μl. The cell suspension mix was loaded on a C1 Single-Cell AutoPrep integrated fluidic circuit (IFC) designed for 10- to 17-µm cells (HT-800, Fluidigm #100-57-80). cDNA synthesis and preamplification was processed following the manufacturer’s instructions (C1 system, Fluidigm). Single cell RNA-sequencing libraries of the cDNA were prepared using Nextera XT DNA library prep kit (Illumina). Libraries were multi-plexed and sequenced according to the manufacturer’s recommendations with paired-end reads using HiSeq2500 platform (Illumina) with an expected depth of 1M reads per single cell, and a final mapping read length of 70bp.
#> GSM3351839 The putative primary somatosensory cortex S1 was microdissected and incubated in 0.05% trypsin at 37°C for 5 minutes. Following tissue digestion, fetal bovine serum was added to the mix and cells were manually dissociated via up-and-down pipetting. Cells were centrifuged 5min at 300g and the pellet was suspended in 1ml of HBSS then passed on a 70µm cell stainer. FT+ cells, gated on the top 5% brightest cells, were finally FAC-sorted on a MoFloAstrios (Beckman). FAC-sorted FT+ cells (18µl) were mixed with the C1 Suspension Reagent (2µl; Fluidigm) yielding a total of 20µl of cell suspension mix with ~500 cells/μl. The cell suspension mix was loaded on a C1 Single-Cell AutoPrep integrated fluidic circuit (IFC) designed for 10- to 17-µm cells (HT-800, Fluidigm #100-57-80). cDNA synthesis and preamplification was processed following the manufacturer’s instructions (C1 system, Fluidigm). Single cell RNA-sequencing libraries of the cDNA were prepared using Nextera XT DNA library prep kit (Illumina). Libraries were multi-plexed and sequenced according to the manufacturer’s recommendations with paired-end reads using HiSeq2500 platform (Illumina) with an expected depth of 1M reads per single cell, and a final mapping read length of 70bp.
#> GSM3351840 The putative primary somatosensory cortex S1 was microdissected and incubated in 0.05% trypsin at 37°C for 5 minutes. Following tissue digestion, fetal bovine serum was added to the mix and cells were manually dissociated via up-and-down pipetting. Cells were centrifuged 5min at 300g and the pellet was suspended in 1ml of HBSS then passed on a 70µm cell stainer. FT+ cells, gated on the top 5% brightest cells, were finally FAC-sorted on a MoFloAstrios (Beckman). FAC-sorted FT+ cells (18µl) were mixed with the C1 Suspension Reagent (2µl; Fluidigm) yielding a total of 20µl of cell suspension mix with ~500 cells/μl. The cell suspension mix was loaded on a C1 Single-Cell AutoPrep integrated fluidic circuit (IFC) designed for 10- to 17-µm cells (HT-800, Fluidigm #100-57-80). cDNA synthesis and preamplification was processed following the manufacturer’s instructions (C1 system, Fluidigm). Single cell RNA-sequencing libraries of the cDNA were prepared using Nextera XT DNA library prep kit (Illumina). Libraries were multi-plexed and sequenced according to the manufacturer’s recommendations with paired-end reads using HiSeq2500 platform (Illumina) with an expected depth of 1M reads per single cell, and a final mapping read length of 70bp.
#> GSM3351841 The putative primary somatosensory cortex S1 was microdissected and incubated in 0.05% trypsin at 37°C for 5 minutes. Following tissue digestion, fetal bovine serum was added to the mix and cells were manually dissociated via up-and-down pipetting. Cells were centrifuged 5min at 300g and the pellet was suspended in 1ml of HBSS then passed on a 70µm cell stainer. FT+ cells, gated on the top 5% brightest cells, were finally FAC-sorted on a MoFloAstrios (Beckman). FAC-sorted FT+ cells (18µl) were mixed with the C1 Suspension Reagent (2µl; Fluidigm) yielding a total of 20µl of cell suspension mix with ~500 cells/μl. The cell suspension mix was loaded on a C1 Single-Cell AutoPrep integrated fluidic circuit (IFC) designed for 10- to 17-µm cells (HT-800, Fluidigm #100-57-80). cDNA synthesis and preamplification was processed following the manufacturer’s instructions (C1 system, Fluidigm). Single cell RNA-sequencing libraries of the cDNA were prepared using Nextera XT DNA library prep kit (Illumina). Libraries were multi-plexed and sequenced according to the manufacturer’s recommendations with paired-end reads using HiSeq2500 platform (Illumina) with an expected depth of 1M reads per single cell, and a final mapping read length of 70bp.
#> GSM3351842 The putative primary somatosensory cortex S1 was microdissected and incubated in 0.05% trypsin at 37°C for 5 minutes. Following tissue digestion, fetal bovine serum was added to the mix and cells were manually dissociated via up-and-down pipetting. Cells were centrifuged 5min at 300g and the pellet was suspended in 1ml of HBSS then passed on a 70µm cell stainer. FT+ cells, gated on the top 5% brightest cells, were finally FAC-sorted on a MoFloAstrios (Beckman). FAC-sorted FT+ cells (18µl) were mixed with the C1 Suspension Reagent (2µl; Fluidigm) yielding a total of 20µl of cell suspension mix with ~500 cells/μl. The cell suspension mix was loaded on a C1 Single-Cell AutoPrep integrated fluidic circuit (IFC) designed for 10- to 17-µm cells (HT-800, Fluidigm #100-57-80). cDNA synthesis and preamplification was processed following the manufacturer’s instructions (C1 system, Fluidigm). Single cell RNA-sequencing libraries of the cDNA were prepared using Nextera XT DNA library prep kit (Illumina). Libraries were multi-plexed and sequenced according to the manufacturer’s recommendations with paired-end reads using HiSeq2500 platform (Illumina) with an expected depth of 1M reads per single cell, and a final mapping read length of 70bp.
#>                          extract_protocol_ch1.1 taxid_ch1
#> GSM3351837 SMARTseq v4 kit (Clontech, # 634888)     10090
#> GSM3351838 SMARTseq v4 kit (Clontech, # 634888)     10090
#> GSM3351839 SMARTseq v4 kit (Clontech, # 634888)     10090
#> GSM3351840 SMARTseq v4 kit (Clontech, # 634888)     10090
#> GSM3351841 SMARTseq v4 kit (Clontech, # 634888)     10090
#> GSM3351842 SMARTseq v4 kit (Clontech, # 634888)     10090
#>                                                                                                                                                                   data_processing
#> GSM3351837 The 5bp UMI sequence (initially sequenced in read1), are appended at the end of the headers of the reads (here the reads we provide are resulting from this operation)
#> GSM3351838 The 5bp UMI sequence (initially sequenced in read1), are appended at the end of the headers of the reads (here the reads we provide are resulting from this operation)
#> GSM3351839 The 5bp UMI sequence (initially sequenced in read1), are appended at the end of the headers of the reads (here the reads we provide are resulting from this operation)
#> GSM3351840 The 5bp UMI sequence (initially sequenced in read1), are appended at the end of the headers of the reads (here the reads we provide are resulting from this operation)
#> GSM3351841 The 5bp UMI sequence (initially sequenced in read1), are appended at the end of the headers of the reads (here the reads we provide are resulting from this operation)
#> GSM3351842 The 5bp UMI sequence (initially sequenced in read1), are appended at the end of the headers of the reads (here the reads we provide are resulting from this operation)
#>                                                                                                  data_processing.1
#> GSM3351837 reads are mapped with tophat v2.0.13 on the mouse genome GRCm38 considering Ensembl78 genes annotations
#> GSM3351838 reads are mapped with tophat v2.0.13 on the mouse genome GRCm38 considering Ensembl78 genes annotations
#> GSM3351839 reads are mapped with tophat v2.0.13 on the mouse genome GRCm38 considering Ensembl78 genes annotations
#> GSM3351840 reads are mapped with tophat v2.0.13 on the mouse genome GRCm38 considering Ensembl78 genes annotations
#> GSM3351841 reads are mapped with tophat v2.0.13 on the mouse genome GRCm38 considering Ensembl78 genes annotations
#> GSM3351842 reads are mapped with tophat v2.0.13 on the mouse genome GRCm38 considering Ensembl78 genes annotations
#>                                                                                  data_processing.2
#> GSM3351837 bam files are processed with umi_tools to deduplicate reads with identical UMI sequence
#> GSM3351838 bam files are processed with umi_tools to deduplicate reads with identical UMI sequence
#> GSM3351839 bam files are processed with umi_tools to deduplicate reads with identical UMI sequence
#> GSM3351840 bam files are processed with umi_tools to deduplicate reads with identical UMI sequence
#> GSM3351841 bam files are processed with umi_tools to deduplicate reads with identical UMI sequence
#> GSM3351842 bam files are processed with umi_tools to deduplicate reads with identical UMI sequence
#>                                                                                                                                                                                           data_processing.3
#> GSM3351837 gene expression are quantifyied using summarizeOverlaps() method of the R package GenomicsAlignments(). Only exonic reads are considered in the quantification and this includes 5' and 3' UTRs.
#> GSM3351838 gene expression are quantifyied using summarizeOverlaps() method of the R package GenomicsAlignments(). Only exonic reads are considered in the quantification and this includes 5' and 3' UTRs.
#> GSM3351839 gene expression are quantifyied using summarizeOverlaps() method of the R package GenomicsAlignments(). Only exonic reads are considered in the quantification and this includes 5' and 3' UTRs.
#> GSM3351840 gene expression are quantifyied using summarizeOverlaps() method of the R package GenomicsAlignments(). Only exonic reads are considered in the quantification and this includes 5' and 3' UTRs.
#> GSM3351841 gene expression are quantifyied using summarizeOverlaps() method of the R package GenomicsAlignments(). Only exonic reads are considered in the quantification and this includes 5' and 3' UTRs.
#> GSM3351842 gene expression are quantifyied using summarizeOverlaps() method of the R package GenomicsAlignments(). Only exonic reads are considered in the quantification and this includes 5' and 3' UTRs.
#>               data_processing.4
#> GSM3351837 Genome_build: GRCm38
#> GSM3351838 Genome_build: GRCm38
#> GSM3351839 Genome_build: GRCm38
#> GSM3351840 Genome_build: GRCm38
#> GSM3351841 Genome_build: GRCm38
#> GSM3351842 Genome_build: GRCm38
#>                                                                                                                                                                                                                                                                                                                  data_processing.5
#> GSM3351837 Supplementary_files_format_and_content: raw_count.tsv.gz contains the result of gene expression quantification by summarizeOverlaps() for each cell. The integers values represent the number of exonic reads mapping into each gene after UMI correction. The first columns contains the gene name that is quantified.
#> GSM3351838 Supplementary_files_format_and_content: raw_count.tsv.gz contains the result of gene expression quantification by summarizeOverlaps() for each cell. The integers values represent the number of exonic reads mapping into each gene after UMI correction. The first columns contains the gene name that is quantified.
#> GSM3351839 Supplementary_files_format_and_content: raw_count.tsv.gz contains the result of gene expression quantification by summarizeOverlaps() for each cell. The integers values represent the number of exonic reads mapping into each gene after UMI correction. The first columns contains the gene name that is quantified.
#> GSM3351840 Supplementary_files_format_and_content: raw_count.tsv.gz contains the result of gene expression quantification by summarizeOverlaps() for each cell. The integers values represent the number of exonic reads mapping into each gene after UMI correction. The first columns contains the gene name that is quantified.
#> GSM3351841 Supplementary_files_format_and_content: raw_count.tsv.gz contains the result of gene expression quantification by summarizeOverlaps() for each cell. The integers values represent the number of exonic reads mapping into each gene after UMI correction. The first columns contains the gene name that is quantified.
#> GSM3351842 Supplementary_files_format_and_content: raw_count.tsv.gz contains the result of gene expression quantification by summarizeOverlaps() for each cell. The integers values represent the number of exonic reads mapping into each gene after UMI correction. The first columns contains the gene name that is quantified.
#>            platform_id   contact_name          contact_email    contact_phone
#> GSM3351837    GPL17021 Julien,,Prados julien.prados@unige.ch +41 22 37 95 396
#> GSM3351838    GPL17021 Julien,,Prados julien.prados@unige.ch +41 22 37 95 396
#> GSM3351839    GPL17021 Julien,,Prados julien.prados@unige.ch +41 22 37 95 396
#> GSM3351840    GPL17021 Julien,,Prados julien.prados@unige.ch +41 22 37 95 396
#> GSM3351841    GPL17021 Julien,,Prados julien.prados@unige.ch +41 22 37 95 396
#> GSM3351842    GPL17021 Julien,,Prados julien.prados@unige.ch +41 22 37 95 396
#>            contact_laboratory contact_department    contact_institute
#> GSM3351837     Denis Jabaudon      Neurosciences University of Geneva
#> GSM3351838     Denis Jabaudon      Neurosciences University of Geneva
#> GSM3351839     Denis Jabaudon      Neurosciences University of Geneva
#> GSM3351840     Denis Jabaudon      Neurosciences University of Geneva
#> GSM3351841     Denis Jabaudon      Neurosciences University of Geneva
#> GSM3351842     Denis Jabaudon      Neurosciences University of Geneva
#>                contact_address contact_city contact_zip/postal_code
#> GSM3351837 Rue Michel Servet 1       Geneva                    1211
#> GSM3351838 Rue Michel Servet 1       Geneva                    1211
#> GSM3351839 Rue Michel Servet 1       Geneva                    1211
#> GSM3351840 Rue Michel Servet 1       Geneva                    1211
#> GSM3351841 Rue Michel Servet 1       Geneva                    1211
#> GSM3351842 Rue Michel Servet 1       Geneva                    1211
#>            contact_country data_row_count    instrument_model library_selection
#> GSM3351837     Switzerland              0 Illumina HiSeq 2500              cDNA
#> GSM3351838     Switzerland              0 Illumina HiSeq 2500              cDNA
#> GSM3351839     Switzerland              0 Illumina HiSeq 2500              cDNA
#> GSM3351840     Switzerland              0 Illumina HiSeq 2500              cDNA
#> GSM3351841     Switzerland              0 Illumina HiSeq 2500              cDNA
#> GSM3351842     Switzerland              0 Illumina HiSeq 2500              cDNA
#>            library_source library_strategy
#> GSM3351837 transcriptomic          RNA-Seq
#> GSM3351838 transcriptomic          RNA-Seq
#> GSM3351839 transcriptomic          RNA-Seq
#> GSM3351840 transcriptomic          RNA-Seq
#> GSM3351841 transcriptomic          RNA-Seq
#> GSM3351842 transcriptomic          RNA-Seq
#>                                                                  relation
#> GSM3351837 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN09901434
#> GSM3351838 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN09901433
#> GSM3351839 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN09901432
#> GSM3351840 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN09901431
#> GSM3351841 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN09901430
#> GSM3351842 BioSample: https://www.ncbi.nlm.nih.gov/biosample/SAMN09901429
#>                                                       relation.1
#> GSM3351837 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX4600063
#> GSM3351838 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX4600064
#> GSM3351839 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX4600065
#> GSM3351840 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX4600066
#> GSM3351841 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX4600067
#> GSM3351842 SRA: https://www.ncbi.nlm.nih.gov/sra?term=SRX4600068
#>            supplementary_file_1 cell.age:ch1 cluster:ch1 num_read:ch1
#> GSM3351837                 NONE          24H    N1d.late       608677
#> GSM3351838                 NONE          24H    N1d.late        68123
#> GSM3351839                 NONE          24H    N1d.late       608394
#> GSM3351840                 NONE          24H    N1d.late       557976
#> GSM3351841                 NONE          24H    N1d.late       121333
#> GSM3351842                 NONE          24H    N1d.late       621786
#>            pct_mt:ch1 pseudo.birthdate.adj.pred:ch1 pseudo.birthdate.pred:ch1
#> GSM3351837       4.3%                         3.411                     0.333
#> GSM3351838       4.8%                         3.434                     0.152
#> GSM3351839       3.5%                         3.963                     0.890
#> GSM3351840       3.4%                         3.316                     0.439
#> GSM3351841       8.7%                         2.687                     0.049
#> GSM3351842       2.9%                         3.553                     0.578
#>            pseudo.birthdate24h:ch1 pseudo.birthdate96h:ch1
#> GSM3351837                   0.417                    <NA>
#> GSM3351838                   0.462                    <NA>
#> GSM3351839                   0.459                    <NA>
#> GSM3351840                   0.505                    <NA>
#> GSM3351841                   0.500                    <NA>
#> GSM3351842                   0.431                    <NA>
#>            pseudo.birthdateprog:ch1 pseudo.differentiation:ch1
#> GSM3351837                     <NA>                      0.409
#> GSM3351838                     <NA>                      0.381
#> GSM3351839                     <NA>                      0.327
#> GSM3351840                     <NA>                      0.378
#> GSM3351841                     <NA>                      0.372
#> GSM3351842                     <NA>                      0.421
#>            pseudo.differentiation.adj.pred:ch1 pseudo.differentiation.pred:ch1
#> GSM3351837                               2.219                           0.592
#> GSM3351838                               1.982                           0.130
#> GSM3351839                               1.982                           0.134
#> GSM3351840                               1.943                           0.119
#> GSM3351841                               1.919                           0.056
#> GSM3351842                               1.908                           0.172
#>            pup.age.at.injection:ch1 tsne1:ch1 tsne2:ch1 pupAGEcellAGE
#> GSM3351837                      E14    -21.73     10.41       E14.24H
#> GSM3351838                      E14    -12.66     -0.39       E14.24H
#> GSM3351839                      E14      4.78      2.06       E14.24H
#> GSM3351840                      E14    -14.95      7.12       E14.24H
#> GSM3351841                      E14    -15.24      7.04       E14.24H
#> GSM3351842                      E14    -19.39     11.40       E14.24H