Assemble Files Based on Component

Assemble individual same-component figures from multiple datasets analyzed and plotted by SJD for cross comparison

Usage

assemblePNG.byComponent(
  SJDScorePlotter.obj,
  component,
  SJD_algorithm,
  group = NA
)

Arguments

SJDScorePlotter.obj

A list outputted by the SJDScorePlotter function

component

numer/order of component of interest to print out, i.e. 1 or c(1,2)

SJD_algorithm

SJD_algorithm name of SJD algorithm, i.e. concatICA

group

group name of the weights group from the SJD_algorithm output, i.e 'Shared.All.13'

Value

A list of images filtered by component

Examples

library(ggplot2)

data(NeuroGenesis4.afterWrap)
data(NeuroGenesis4.info)

SampleMetaNamesTable = data.frame(
   row.names = names(NeuroGenesis4),
   Type = c('Yaxis','Yaxis','2Dscatter','2Dscatter'),
   XaxisColumn = c("X","DAYx","tSNE_1","tsne1:ch1"),
   YaxisColumn = c("PJDscores","PJDscores","tSNE_2","tsne2:ch1"),
   COLaxisColumn = c("color","colorBYlabelsX","PJDscores","PJDscores"),
   PCHColumn = c("","","","")
)

grp = list(
Shared.All.4 = c(1 : 4),
Shared.bulk.2 = c(1, 2),
Shared.sc.2 = c(3, 4),
Hs.Meisnr.1 = c(1),
Hs.AZ.1 = c(2),
Gesch.1 = c(3),
Telley.1 = c(4)
)
dims = c(2, 2, 2, 2, 2, 2, 2)

lbb = "NeuroGenesis4.p2"

twoStageLCA.out = twoStageLCA(dataset = NeuroGenesis4.afterWrap, group = grp, comp_num = dims)

SJDScorePlotter.obj = SJDScorePlotter(
    SJDalg = "twoStageLCA",
    scores = twoStageLCA.out$score_list,
    lbb = lbb,
    info = NeuroGenesis4.info,
    SampleMetaNamesTable = SampleMetaNamesTable
)
#> [1] "**********************************************************"
#> [1] "plotting scores for twoStageLCA anaysis."
#> [1] "*******************************"
#> [1] "dataset: Meissner.inVitro.bulk.Hs"
#> [1] "grouping: Shared.All.4"
#> [1] 1
#> [1] 2
#> [1] "grouping: Shared.bulk.2"
#> [1] 1
#> [1] 2
#> [1] "grouping: Shared.sc.2"
#> [1] "No SJD scores for grouping : Shared.sc.2, moving to the next SJD output for Meissner.inVitro.bulk.Hs"
#> [1] "grouping: Hs.Meisnr.1"
#> [1] 1
#> [1] 2
#> [1] "grouping: Hs.AZ.1"
#> [1] "No SJD scores for grouping : Hs.AZ.1, moving to the next SJD output for Meissner.inVitro.bulk.Hs"
#> [1] "grouping: Gesch.1"
#> [1] "No SJD scores for grouping : Gesch.1, moving to the next SJD output for Meissner.inVitro.bulk.Hs"
#> [1] "grouping: Telley.1"
#> [1] "No SJD scores for grouping : Telley.1, moving to the next SJD output for Meissner.inVitro.bulk.Hs"
#> [1] "*******************************"
#> [1] "dataset: LIBD.AZ.inVitro.bulk.Hs"
#> [1] "grouping: Shared.All.4"
#> [1] 1
#> [1] 2
#> [1] "grouping: Shared.bulk.2"
#> [1] 1
#> [1] 2
#> [1] "grouping: Shared.sc.2"
#> [1] "No SJD scores for grouping : Shared.sc.2, moving to the next SJD output for LIBD.AZ.inVitro.bulk.Hs"
#> [1] "grouping: Hs.Meisnr.1"
#> [1] "No SJD scores for grouping : Hs.Meisnr.1, moving to the next SJD output for LIBD.AZ.inVitro.bulk.Hs"
#> [1] "grouping: Hs.AZ.1"
#> [1] 1
#> [1] 2
#> [1] "grouping: Gesch.1"
#> [1] "No SJD scores for grouping : Gesch.1, moving to the next SJD output for LIBD.AZ.inVitro.bulk.Hs"
#> [1] "grouping: Telley.1"
#> [1] "No SJD scores for grouping : Telley.1, moving to the next SJD output for LIBD.AZ.inVitro.bulk.Hs"
#> [1] "*******************************"
#> [1] "dataset: Geschwind.inVivo.sc.Hs"
#> [1] "grouping: Shared.All.4"
#> [1] 1
#> [1] 2
#> [1] "grouping: Shared.bulk.2"
#> [1] "No SJD scores for grouping : Shared.bulk.2, moving to the next SJD output for Geschwind.inVivo.sc.Hs"
#> [1] "grouping: Shared.sc.2"
#> [1] 1
#> [1] 2
#> [1] "grouping: Hs.Meisnr.1"
#> [1] "No SJD scores for grouping : Hs.Meisnr.1, moving to the next SJD output for Geschwind.inVivo.sc.Hs"
#> [1] "grouping: Hs.AZ.1"
#> [1] "No SJD scores for grouping : Hs.AZ.1, moving to the next SJD output for Geschwind.inVivo.sc.Hs"
#> [1] "grouping: Gesch.1"
#> [1] 1
#> [1] 2
#> [1] "grouping: Telley.1"
#> [1] "No SJD scores for grouping : Telley.1, moving to the next SJD output for Geschwind.inVivo.sc.Hs"
#> [1] "*******************************"
#> [1] "dataset: Jabaudon.inVivo.sc.Mm"
#> [1] "grouping: Shared.All.4"
#> [1] 1
#> [1] 2
#> [1] "grouping: Shared.bulk.2"
#> [1] "No SJD scores for grouping : Shared.bulk.2, moving to the next SJD output for Jabaudon.inVivo.sc.Mm"
#> [1] "grouping: Shared.sc.2"
#> [1] 1
#> [1] 2
#> [1] "grouping: Hs.Meisnr.1"
#> [1] "No SJD scores for grouping : Hs.Meisnr.1, moving to the next SJD output for Jabaudon.inVivo.sc.Mm"
#> [1] "grouping: Hs.AZ.1"
#> [1] "No SJD scores for grouping : Hs.AZ.1, moving to the next SJD output for Jabaudon.inVivo.sc.Mm"
#> [1] "grouping: Gesch.1"
#> [1] "No SJD scores for grouping : Gesch.1, moving to the next SJD output for Jabaudon.inVivo.sc.Mm"
#> [1] "grouping: Telley.1"
#> [1] 1
#> [1] 2

assemble.byComponent.obj = assemblePNG.byComponent(
SJDScorePlotter.obj = SJDScorePlotter.obj,
component = c(1, 2),
SJD_algorithm = "twoStageLCA",
group = 'Shared.All.4')