Percentage of Variance Explained for separate data set

Compute the PVE (percentage of variance explained) for each data set

Usage

pveSep(dataset, list_score, list_component)

Arguments

dataset

A list of data sets for input

list_score

A list of extracted scores by the corresponding algorithm

list_component

A list of components comptuted by the corresponding algorithm

Value

The list of scores

Examples

dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
comp_num = 2
res_sepPCA = sepPCA(dataset, comp_num)
pveSep(dataset, res_sepPCA$score_list, res_sepPCA$linked_component_list)
#> $dataset_No.1
#>                                                      subject_No.1 subject_No.2
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174 -0.003328291  0.001637634
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  0.005895471  0.011104309
#>                                                      subject_No.3  subject_No.4
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  0.005354008  0.0061163104
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  0.007347788 -0.0005548611
#>                                                      subject_No.5  subject_No.6
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  0.003047467 -0.0006000776
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483 -0.002038949 -0.0062102478
#>                                                       subject_No.7 subject_No.8
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174 -0.0002377363 0.0018828480
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  0.0006400535 0.0003632946
#>                                                      subject_No.9 subject_No.10
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  0.001563414   0.007124862
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483 -0.002532921  -0.005469364
#>                                                      subject_No.11
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174   0.007506413
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.004889506
#>                                                      subject_No.12
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.002653039
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.001055445
#>                                                      subject_No.13
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.001000789
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.002331137
#>                                                      subject_No.14
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174   0.001509456
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483   0.003198024
#>                                                      subject_No.15
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.002954268
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.003547815
#>                                                      subject_No.16
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174 -0.0031761234
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483 -0.0003604353
#>                                                      subject_No.17
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174 -0.0013415453
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483 -0.0005968003
#>                                                      subject_No.18
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.001260152
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483   0.006825200
#>                                                      subject_No.19
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174   0.006006984
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.003880804
#>                                                      subject_No.20
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174   0.008236343
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.002075164
#>                                                      subject_No.21
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174   0.001247889
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483   0.005359478
#>                                                      subject_No.22
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.007921062
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.001410534
#>                                                      subject_No.23
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174   0.002988066
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.002215995
#>                                                      subject_No.24
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.004688390
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.001027427
#>                                                      subject_No.25
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174   0.007399843
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483   0.004282864
#>                                                      subject_No.26
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  0.0055435060
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483 -0.0003195842
#>                                                      subject_No.27
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174 -0.0053809860
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483 -0.0008463669
#>                                                      subject_No.28
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.005175293
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.004107678
#>                                                      subject_No.29
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.004104925
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.007689743
#>                                                      subject_No.30
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.001114217
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483   0.003674191
#>                                                      subject_No.31
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174   0.003041121
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483   0.006025926
#>                                                      subject_No.32
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174   0.008347782
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.001534792
#>                                                      subject_No.33
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174   0.001834988
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.006872969
#>                                                      subject_No.34
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.002248361
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483   0.002050168
#>                                                      subject_No.35
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.003064286
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.004294616
#>                                                      subject_No.36
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.007599815
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483   0.003138609
#>                                                      subject_No.37
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.002319395
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.001971029
#>                                                      subject_No.38
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.001205233
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483   0.010215037
#>                                                      subject_No.39
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.007595254
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.000968588
#>                                                      subject_No.40
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174   0.003566083
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.006118690
#>                                                      subject_No.41
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174   0.004464625
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483   0.005634366
#>                                                      subject_No.42
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  0.0004239954
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  0.0004938730
#>                                                      subject_No.43
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.004831121
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483   0.003858843
#>                                                      subject_No.44
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.001748605
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.005148402
#>                                                      subject_No.45
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.003022400
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.003015066
#>                                                      subject_No.46
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.007823090
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483  -0.001212101
#>                                                      subject_No.47
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.001645625
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483   0.005464428
#>                                                      subject_No.48
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  0.0022809591
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483 -0.0008739412
#>                                                      subject_No.49
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  -0.003088238
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483   0.001389531
#>                                                      subject_No.50
#> dataset_No.1_subcomp.1, PVE: 0.051918, PVE: 0.053174  3.719042e-06
#> dataset_No.1_subcomp.2, PVE: 0.050949, PVE: 0.008483 -1.790482e-03
#>