SJD Wrap

wrapping up expression matrices (of different species) with only shared genes as SJD input

sjdWrap(
  data.list,
  species.vector,
  geneType.vector,
  geneType.out = "symbol",
  species.out
)

Arguments

data.list
input list of expression matrices from different species i.e human and mouse datasets

species.vector
character of species type of each matrix i.e c('human', 'mouse', 'mouse')

geneType.vector
character of gene/rowname type of each matrix i.e c("symbol","ensembl","symbol")

geneType.out
character of output gene/rowname type of each matrix i.e "symbol"

species.out
character of output species type for gene/rowname

Value

a list of expression matrices (of different species) with only shared genes and

Examples

data(NeuroGenesis4)
SJDdataIN = sjdWrap(
data.list = NeuroGenesis4,
species.vector=c("human","human","human","mouse"),
geneType.vector=c("symbol","ensembl","symbol","symbol"),
geneType.out="symbol",
species.out="human")
#> Using biomaRt to connect gene IDs across 4 datasets:
#> Getting biomaRt IDs for dataset 1 
#> You have input  40  genes
#> We found  40  matches
#> 1  of those are duplicates and only keeping the 1st of each
#> Getting biomaRt IDs for dataset 2 
#> You have input  40  genes
#> We found  37  matches
#> 1  of those are duplicates and only keeping the 1st of each
#> Getting biomaRt IDs for dataset 3 
#> You have input  40  genes
#> We found  39  matches
#> 0  of those are duplicates and only keeping the 1st of each
#> Getting biomaRt IDs for dataset 4 
#> You have input  40  genes
#> We found  32  matches
#> 1  of those are duplicates and only keeping the 1st of each
#> constructed 4 tables of cross-species matching genes
#> we found 25 shared genes in 4 datasets
#> new data list of 4 datasets constructed