Two-staged Linked Component Analysis

Two-staged Linked Component Analysis

Usage

twoStageLCA(
  dataset,
  group,
  comp_num,
  weighting = NULL,
  backup = 0,
  plotting = FALSE,
  proj_dataset = NULL,
  proj_group = NULL,
  enable_normalization = TRUE,
  column_sum_normalization = TRUE,
  screen_prob = NULL
)

Arguments

dataset

A list of dataset to be analyzed

group

A list of grouping of the datasets, indicating the relationship between datasets

comp_num

A vector indicates the dimension of each compoent

weighting

Weighting of each dataset, initialized to be NULL

backup

A positive scalar to determine how many PCs to over select

plotting

A boolean value to determine whether to plot the scree plot or not, default to be False

proj_dataset

The datasets to be projected on

proj_group

The grouping of projected data sets

enable_normalization

An argument to decide whether to use normalizaiton or not, default is TRUE

column_sum_normalization

An argument to decide whether to use column sum normalization or not, default it TRUE

screen_prob

A vector of probabilies for genes to be chosen

Value

A list contains the component and the score of each dataset on every component after 2sLCA algorithm

Examples

dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50),
matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
group = list(c(1, 2, 3, 4), c(1, 2), c(3, 4), c(1, 3), c(2, 4), c(1), c(2), c(3), c(4))
comp_num = c(2, 2, 2, 2, 2, 2, 2, 2, 2)
proj_dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
proj_group = list(c(TRUE, TRUE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, FALSE))
res_twoStageLCA = twoStageLCA(
dataset,
group,
comp_num,
proj_dataset = proj_dataset,
proj_group = proj_group)