Percentage of Variance Explained for separate data set

Compute the PVE (percentage of variance explained) for each data set

pveSep(dataset, list_score, list_component)

Arguments

dataset
A list of data sets for input

list_score
A list of extracted scores by the corresponding algorithm

list_component
A list of components comptuted by the corresponding algorithm

Value

The list of scores

Examples

dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
comp_num = 2
res_sepPCA = sepPCA(dataset, comp_num)
pveSep(dataset, res_sepPCA$score_list, res_sepPCA$linked_component_list)
#> $dataset_No.1
#>                                                      subject_No.1 subject_No.2
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810    2.3629466    -2.801871
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807    0.6499845     1.263631
#>                                                      subject_No.3 subject_No.4
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810    -2.051842    0.1203612
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     3.245521    0.5892438
#>                                                      subject_No.5 subject_No.6
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810   -0.2274785    -1.465924
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807   -4.7128204    -1.449247
#>                                                      subject_No.7 subject_No.8
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     3.028705    2.9360805
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807    -2.196647   -0.6623563
#>                                                      subject_No.9 subject_No.10
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810  -1.74197410      2.643085
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807  -0.03288524     -3.120455
#>                                                      subject_No.11
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810   -4.26347882
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807    0.07356096
#>                                                      subject_No.12
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810   -0.72747745
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807   -0.04338173
#>                                                      subject_No.13
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810   0.002716421
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807   0.103423204
#>                                                      subject_No.14
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     -1.395175
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     -1.147377
#>                                                      subject_No.15
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810   -2.00778905
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807   -0.02218161
#>                                                      subject_No.16
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810    -1.5257966
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     0.8090099
#>                                                      subject_No.17
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     -2.469410
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807      3.512999
#>                                                      subject_No.18
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     0.1769953
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807    -0.3352157
#>                                                      subject_No.19
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     0.5963105
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807    -2.0489367
#>                                                      subject_No.20
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810      2.265825
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807      3.378187
#>                                                      subject_No.21
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     -3.325468
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     -1.044854
#>                                                      subject_No.22
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     3.9252674
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807    -0.3035396
#>                                                      subject_No.23
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     0.6982377
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807    -3.0860843
#>                                                      subject_No.24
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     -2.816837
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     -3.655045
#>                                                      subject_No.25
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810    -0.6826798
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     2.0324311
#>                                                      subject_No.26
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810    -0.2682604
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807    -0.7551556
#>                                                      subject_No.27
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     -4.317708
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     -4.795504
#>                                                      subject_No.28
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810      1.723269
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807      1.368315
#>                                                      subject_No.29
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810    -0.7001556
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     0.8012285
#>                                                      subject_No.30
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810   -0.07254962
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807    3.13405671
#>                                                      subject_No.31
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     0.3728344
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     0.1649489
#>                                                      subject_No.32
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     5.4747313
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     0.9404153
#>                                                      subject_No.33
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     0.9718157
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     1.9816486
#>                                                      subject_No.34
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810      3.519333
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807      3.049038
#>                                                      subject_No.35
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810    -0.7223053
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807    -0.8616084
#>                                                      subject_No.36
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810    -0.5838016
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     1.6373256
#>                                                      subject_No.37
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     -3.596616
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807      1.676363
#>                                                      subject_No.38
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810      2.746376
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807      0.369335
#>                                                      subject_No.39
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     2.5207879
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     0.2679324
#>                                                      subject_No.40
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810      1.523641
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807      1.601613
#>                                                      subject_No.41
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810      2.047667
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807      1.314036
#>                                                      subject_No.42
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     -3.309825
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807      2.618636
#>                                                      subject_No.43
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810      1.539551
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     -5.559850
#>                                                      subject_No.44
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     -3.500759
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     -2.152748
#>                                                      subject_No.45
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810      1.671094
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     -3.162102
#>                                                      subject_No.46
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810      1.970054
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     -3.202369
#>                                                      subject_No.47
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     -1.222358
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807      2.888377
#>                                                      subject_No.48
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     1.7210491
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     0.3994721
#>                                                      subject_No.49
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810     -0.106310
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807      2.695126
#>                                                      subject_No.50
#> dataset_No.1_subcomp.1, PVE: 0.052231, PVE: 0.001810    -0.5742701
#> dataset_No.1_subcomp.2, PVE: 0.051135, PVE: 0.001807     1.8600408
#>