Percentage of Variance Explained for separate data set

Compute the PVE (percentage of variance explained) for each data set

Usage

pveSep(dataset, list_score, list_component)

Arguments

dataset

A list of data sets for input

list_score

A list of extracted scores by the corresponding algorithm

list_component

A list of components comptuted by the corresponding algorithm

Value

The list of scores

Examples

dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
comp_num = 2
res_sepPCA = sepPCA(dataset, comp_num)
pveSep(dataset, res_sepPCA$score_list, res_sepPCA$linked_component_list)
#> $dataset_No.1
#>                                                      subject_No.1 subject_No.2
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  0.005170197  0.003277219
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121 -0.000451882 -0.007273454
#>                                                      subject_No.3 subject_No.4
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  0.003973102 0.0002117434
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  0.006428405 0.0062501574
#>                                                      subject_No.5  subject_No.6
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597 -0.001760661 -0.0014172771
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121 -0.005391612  0.0002074341
#>                                                       subject_No.7 subject_No.8
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597 -0.0003876718  0.006638273
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121 -0.0022856258 -0.002427173
#>                                                      subject_No.9 subject_No.10
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597 0.0039228912  -0.010978815
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121 0.0001950275  -0.000447776
#>                                                      subject_No.11
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597 -0.0005437079
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121 -0.0036800591
#>                                                      subject_No.12
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  0.0045395108
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121 -0.0002724696
#>                                                      subject_No.13
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597 -0.0031847964
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121 -0.0009354165
#>                                                      subject_No.14
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  0.0004282624
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  0.0078622096
#>                                                      subject_No.15
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597   0.002814451
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  -0.001721630
#>                                                      subject_No.16
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597   0.004115676
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  -0.008044665
#>                                                      subject_No.17
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597   0.002927328
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121   0.009801242
#>                                                      subject_No.18
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  0.0002524831
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  0.0007927868
#>                                                      subject_No.19
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597   0.004444059
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  -0.003220107
#>                                                      subject_No.20
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597   0.004078303
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121   0.004036384
#>                                                      subject_No.21
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597   0.006302836
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121   0.002085125
#>                                                      subject_No.22
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597   0.004053028
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  -0.007484716
#>                                                      subject_No.23
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597   0.005875607
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121   0.003669864
#>                                                      subject_No.24
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597   0.004562340
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121   0.006304561
#>                                                      subject_No.25
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  -0.003657365
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  -0.001340592
#>                                                      subject_No.26
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597   0.003345328
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121   0.008809878
#>                                                      subject_No.27
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  -0.007951284
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  -0.002447639
#>                                                      subject_No.28
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597 -0.0001135516
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121 -0.0001942014
#>                                                      subject_No.29
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597 -0.0069157940
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121 -0.0009761249
#>                                                      subject_No.30
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  0.0011715706
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  0.0004507421
#>                                                      subject_No.31
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  -0.005059233
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121   0.002073347
#>                                                      subject_No.32
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  -0.000982777
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121   0.001001296
#>                                                      subject_No.33
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597   0.003839120
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  -0.002813761
#>                                                      subject_No.34
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  -0.010880814
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121   0.005044448
#>                                                      subject_No.35
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597 -0.0006372848
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121 -0.0071676972
#>                                                      subject_No.36
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  -0.003297835
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  -0.002251017
#>                                                      subject_No.37
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597   0.001355491
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121   0.007027169
#>                                                      subject_No.38
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  -0.005960817
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121   0.006080517
#>                                                      subject_No.39
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597   0.005484740
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  -0.000694527
#>                                                      subject_No.40
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  0.0001871763
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121 -0.0005264952
#>                                                      subject_No.41
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  -0.008362175
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121   0.000189674
#>                                                      subject_No.42
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  -0.002624562
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  -0.006576830
#>                                                      subject_No.43
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  0.0005270331
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  0.0046117429
#>                                                      subject_No.44
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  -0.001578349
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  -0.006457342
#>                                                      subject_No.45
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  -0.002094095
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  -0.004210145
#>                                                      subject_No.46
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597 -0.0053050775
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  0.0001465597
#>                                                      subject_No.47
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597   0.004207142
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  -0.001641287
#>                                                      subject_No.48
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597   0.003279348
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  -0.004937337
#>                                                      subject_No.49
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597 -0.0027212857
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121  0.0002992554
#>                                                      subject_No.50
#> dataset_No.1_subcomp.1, PVE: 0.054465, PVE: 0.005597  -0.004569029
#> dataset_No.1_subcomp.2, PVE: 0.053681, PVE: 0.016121   0.002503756
#>