Percentage of Variance Explained for separate data set

Compute the PVE (percentage of variance explained) for each data set

pveSep(dataset, list_score, list_component)

Arguments

dataset
A list of data sets for input

list_score
A list of extracted scores by the corresponding algorithm

list_component
A list of components comptuted by the corresponding algorithm

Value

The list of scores

Examples

dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
comp_num = 2
res_sepPCA = sepPCA(dataset, comp_num)
pveSep(dataset, res_sepPCA$score_list, res_sepPCA$linked_component_list)
#> $dataset_No.1
#>                                                      subject_No.1 subject_No.2
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860   2.44623686    -2.056660
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803  -0.06653659     1.726016
#>                                                      subject_No.3 subject_No.4
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860   -0.3839052     1.418653
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803    3.3355231    -0.723363
#>                                                      subject_No.5 subject_No.6
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860    -1.310046   -1.6986930
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803    -4.236779   -0.9823784
#>                                                      subject_No.7 subject_No.8
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     1.053832     3.204591
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803    -2.520867    -1.021916
#>                                                      subject_No.9 subject_No.10
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860   -0.1354992      2.112767
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803    0.6176741     -4.014864
#>                                                      subject_No.11
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     -3.367827
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803      0.307349
#>                                                      subject_No.12
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     -1.075778
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803      1.617496
#>                                                      subject_No.13
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     0.4627409
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803    -1.0192675
#>                                                      subject_No.14
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860    -2.0741197
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     0.2494488
#>                                                      subject_No.15
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860    -2.0822299
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     0.4497253
#>                                                      subject_No.16
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     -1.583931
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803      2.194370
#>                                                      subject_No.17
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860    -0.1046751
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     3.1946952
#>                                                      subject_No.18
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     0.1074933
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803    -0.1404233
#>                                                      subject_No.19
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860    0.01625912
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803   -1.51710861
#>                                                      subject_No.20
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860      2.842454
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803      2.953893
#>                                                      subject_No.21
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860   -3.26252764
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803   -0.09438407
#>                                                      subject_No.22
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860      3.541787
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     -1.644199
#>                                                      subject_No.23
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860    -0.3705677
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803    -3.2560937
#>                                                      subject_No.24
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     -4.326112
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     -2.440696
#>                                                      subject_No.25
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860    -0.5324596
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     2.4318460
#>                                                      subject_No.26
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860    -0.8234929
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803    -0.7770536
#>                                                      subject_No.27
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     -5.937508
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     -3.070959
#>                                                      subject_No.28
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860      1.472513
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803      1.407555
#>                                                      subject_No.29
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     0.3921077
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     0.9808792
#>                                                      subject_No.30
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860      1.437302
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803      2.714385
#>                                                      subject_No.31
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860    -0.5178884
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     0.4057926
#>                                                      subject_No.32
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860      4.602982
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     -1.488727
#>                                                      subject_No.33
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     0.2146758
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     2.4681466
#>                                                      subject_No.34
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860      3.632051
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803      2.295937
#>                                                      subject_No.35
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860    -0.2939136
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803    -1.1109982
#>                                                      subject_No.36
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860   -0.07870569
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803    1.24031455
#>                                                      subject_No.37
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     -2.901201
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803      3.188690
#>                                                      subject_No.38
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     3.3741828
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803    -0.9780534
#>                                                      subject_No.39
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860      3.509276
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     -1.031999
#>                                                      subject_No.40
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     2.9125744
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     0.9329575
#>                                                      subject_No.41
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     3.0867766
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     0.5294393
#>                                                      subject_No.42
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     -2.377716
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803      2.986981
#>                                                      subject_No.43
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860    -0.5285799
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803    -5.7779605
#>                                                      subject_No.44
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860    -4.5796132
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803    -0.3855031
#>                                                      subject_No.45
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     0.3276572
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803    -3.8201570
#>                                                      subject_No.46
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860     -0.116201
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     -3.575888
#>                                                      subject_No.47
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860    -0.8030144
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     2.6794685
#>                                                      subject_No.48
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860    1.43058916
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803   -0.02464731
#>                                                      subject_No.49
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860   -0.04826947
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803    2.69961012
#>                                                      subject_No.50
#> dataset_No.1_subcomp.1, PVE: 0.053039, PVE: 0.001860    -0.6794442
#> dataset_No.1_subcomp.2, PVE: 0.050946, PVE: 0.001803     2.2271495
#>