Percentage of Variance Explained for separate data set

Compute the PVE (percentage of variance explained) for each data set

Usage

pveSep(dataset, list_score, list_component)

Arguments

dataset

A list of data sets for input

list_score

A list of extracted scores by the corresponding algorithm

list_component

A list of components comptuted by the corresponding algorithm

Value

The list of scores

Examples

dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
comp_num = 2
res_sepPCA = sepPCA(dataset, comp_num)
pveSep(dataset, res_sepPCA$score_list, res_sepPCA$linked_component_list)
#> $dataset_No.1
#>                                                       subject_No.1 subject_No.2
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504 -0.0086358124 -0.006864778
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  0.0006770942  0.005766750
#>                                                      subject_No.3 subject_No.4
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  0.005097892  0.010430439
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  0.004892070 -0.006778312
#>                                                      subject_No.5 subject_No.6
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504 -0.001400482 -0.006485789
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794 -0.001409331  0.004531620
#>                                                      subject_No.7 subject_No.8
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504 -0.003201511  0.004597578
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  0.007858740 -0.001079773
#>                                                       subject_No.9
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504 -0.0091366739
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  0.0002500813
#>                                                      subject_No.10
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  -0.002935509
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  -0.004928175
#>                                                      subject_No.11
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  -0.002834196
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794   0.002178115
#>                                                      subject_No.12
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  -0.004011823
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794   0.003331101
#>                                                      subject_No.13
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  0.0057931115
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  0.0002328272
#>                                                      subject_No.14
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  -0.006072798
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  -0.006268773
#>                                                      subject_No.15
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504 -0.0006650488
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794 -0.0024186252
#>                                                      subject_No.16
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  0.0077750480
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794 -0.0009176174
#>                                                      subject_No.17
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  -0.002177990
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  -0.003536764
#>                                                      subject_No.18
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504 -0.0002671811
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  0.0054848967
#>                                                      subject_No.19
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  -0.002689829
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794   0.001025270
#>                                                      subject_No.20
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504   0.001360404
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  -0.009333374
#>                                                      subject_No.21
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  -0.005582687
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  -0.004104045
#>                                                      subject_No.22
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  -0.005911710
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  -0.005432824
#>                                                      subject_No.23
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504 -0.0006975859
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  0.0028175346
#>                                                      subject_No.24
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504 -0.0002081154
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794 -0.0007554597
#>                                                      subject_No.25
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504   0.004283422
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  -0.007206840
#>                                                      subject_No.26
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504   0.004967553
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  -0.001967268
#>                                                      subject_No.27
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504   0.001337641
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  -0.001749838
#>                                                      subject_No.28
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  0.0005088883
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794 -0.0018744563
#>                                                      subject_No.29
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504   0.002614464
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794   0.003212094
#>                                                      subject_No.30
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504   0.003392329
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794   0.008456545
#>                                                      subject_No.31
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  0.0006248878
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  0.0016333105
#>                                                      subject_No.32
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  0.0001160368
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794 -0.0073933928
#>                                                      subject_No.33
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  1.016460e-02
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  5.528957e-05
#>                                                      subject_No.34
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  -0.001056928
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  -0.004238054
#>                                                      subject_No.35
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  0.0064824904
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  0.0005040804
#>                                                      subject_No.36
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504   0.001195681
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  -0.002101193
#>                                                      subject_No.37
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504   0.008450907
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794   0.003871364
#>                                                      subject_No.38
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504   0.003449056
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  -0.000212370
#>                                                      subject_No.39
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504   0.004105610
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794   0.004868686
#>                                                      subject_No.40
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  -0.006542368
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  -0.002044269
#>                                                      subject_No.41
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504 -0.0022351086
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794 -0.0004204626
#>                                                      subject_No.42
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  -0.001595403
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  -0.001741886
#>                                                      subject_No.43
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504   0.001345617
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794   0.003132625
#>                                                      subject_No.44
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  -0.002362031
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794   0.002623615
#>                                                      subject_No.45
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  -0.001093286
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  -0.009955164
#>                                                      subject_No.46
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  0.0001715464
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  0.0015203980
#>                                                      subject_No.47
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  -0.002180467
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794   0.005502151
#>                                                      subject_No.48
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  0.0004132554
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794  0.0022408512
#>                                                      subject_No.49
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504   0.004414121
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794   0.007682843
#>                                                      subject_No.50
#> dataset_No.1_subcomp.1, PVE: 0.059074, PVE: 0.004504  -0.006247464
#> dataset_No.1_subcomp.2, PVE: 0.051610, PVE: 0.010794   0.003518312
#>