Percentage of Variance Explained for separate data set

Compute the PVE (percentage of variance explained) for each data set

Usage

pveSep(dataset, list_score, list_component)

Arguments

dataset

A list of data sets for input

list_score

A list of extracted scores by the corresponding algorithm

list_component

A list of components comptuted by the corresponding algorithm

Value

The list of scores

Examples

dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
comp_num = 2
res_sepPCA = sepPCA(dataset, comp_num)
pveSep(dataset, res_sepPCA$score_list, res_sepPCA$linked_component_list)
#> $dataset_No.1
#>                                                      subject_No.1  subject_No.2
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  0.005795634  0.0008567605
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  0.001473243 -0.0018556096
#>                                                      subject_No.3 subject_No.4
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437 -0.007343093  0.003270871
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574 -0.001737167 -0.002320184
#>                                                       subject_No.5 subject_No.6
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437 -0.0094370831  0.001136580
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  0.0008491078  0.008403351
#>                                                       subject_No.7 subject_No.8
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437 -0.0006843763  0.004515469
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  0.0022032448  0.007912714
#>                                                      subject_No.9 subject_No.10
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  0.003419818  -0.009637427
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  0.008071769  -0.001083783
#>                                                      subject_No.11
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437   0.003530809
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574   0.004122503
#>                                                      subject_No.12
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  0.0053968580
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  0.0003792852
#>                                                      subject_No.13
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  0.0059776239
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  0.0001469307
#>                                                      subject_No.14
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437   0.003152558
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  -0.003322997
#>                                                      subject_No.15
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437 -0.0003982072
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574 -0.0097145429
#>                                                      subject_No.16
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437   0.006730505
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574   0.003595457
#>                                                      subject_No.17
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  -0.001224536
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574   0.003237734
#>                                                      subject_No.18
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437   0.001471974
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  -0.006916428
#>                                                      subject_No.19
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437 -0.0020273373
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  0.0004162501
#>                                                      subject_No.20
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  -0.003967424
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  -0.002046359
#>                                                      subject_No.21
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  0.0004031584
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  0.0032644078
#>                                                      subject_No.22
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437 -0.0007776945
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  0.0027206818
#>                                                      subject_No.23
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437   0.001691088
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574   0.004295099
#>                                                      subject_No.24
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437   0.002745351
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  -0.005530389
#>                                                      subject_No.25
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  6.347119e-05
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574 -3.577367e-03
#>                                                      subject_No.26
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437   0.001124719
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  -0.001594598
#>                                                      subject_No.27
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437   0.002511965
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574   0.003863212
#>                                                      subject_No.28
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437   0.001701057
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  -0.007985903
#>                                                      subject_No.29
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  -0.008982659
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574   0.004319102
#>                                                      subject_No.30
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437 -0.0084614109
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  0.0002905205
#>                                                      subject_No.31
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  0.0003304152
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574 -0.0034099319
#>                                                      subject_No.32
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437   0.003053615
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  -0.002236053
#>                                                      subject_No.33
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437 -0.0000749015
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  0.0042003082
#>                                                      subject_No.34
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  -0.002474336
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574   0.007025532
#>                                                      subject_No.35
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437   0.006521387
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  -0.002745434
#>                                                      subject_No.36
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437 -0.0010593961
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  0.0004897315
#>                                                      subject_No.37
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437   0.004638522
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  -0.002486691
#>                                                      subject_No.38
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  -0.003328098
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  -0.007612977
#>                                                      subject_No.39
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437   0.012200281
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  -0.001047048
#>                                                      subject_No.40
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  -0.006404947
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574   0.006458537
#>                                                      subject_No.41
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  -0.004451613
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  -0.002290200
#>                                                      subject_No.42
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  -0.004398538
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574   0.002042626
#>                                                      subject_No.43
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  -0.000292194
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574   0.006829214
#>                                                      subject_No.44
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437 -0.0003471153
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574 -0.0037415023
#>                                                      subject_No.45
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437 -0.0040705709
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574 -0.0003800411
#>                                                      subject_No.46
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437 -0.0002551777
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574 -0.0011110011
#>                                                      subject_No.47
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  -0.002523096
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  -0.008599809
#>                                                      subject_No.48
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  0.0008950106
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574  0.0009435563
#>                                                      subject_No.49
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437 -0.0008779861
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574 -0.0022605339
#>                                                      subject_No.50
#> dataset_No.1_subcomp.1, PVE: 0.053892, PVE: 0.003437  0.0003637159
#> dataset_No.1_subcomp.2, PVE: 0.051721, PVE: 0.007574 -0.0019475699
#>