Percentage of Variance Explained for separate data set

Compute the PVE (percentage of variance explained) for each data set

Usage

pveSep(dataset, list_score, list_component)

Arguments

dataset

A list of data sets for input

list_score

A list of extracted scores by the corresponding algorithm

list_component

A list of components comptuted by the corresponding algorithm

Value

The list of scores

Examples

dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
comp_num = 2
res_sepPCA = sepPCA(dataset, comp_num)
pveSep(dataset, res_sepPCA$score_list, res_sepPCA$linked_component_list)
#> $dataset_No.1
#>                                                      subject_No.1 subject_No.2
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     1.269362    0.8822844
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643    -1.020803    2.0057512
#>                                                      subject_No.3 subject_No.4
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872    0.4119452   -1.2206348
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643   -4.1562911    0.4497719
#>                                                      subject_No.5 subject_No.6
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872   -0.5999940    -1.627559
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643    0.1316427     1.396466
#>                                                      subject_No.7 subject_No.8
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872    -2.858345    -1.116851
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     2.177357     1.272958
#>                                                      subject_No.9 subject_No.10
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872    0.8852413     -3.566629
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643    1.1510677      2.985520
#>                                                      subject_No.11
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872    -0.6313001
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     4.5903746
#>                                                      subject_No.12
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     1.3058714
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643    -0.6927287
#>                                                      subject_No.13
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872      4.955926
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     -3.333619
#>                                                      subject_No.14
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872      3.853117
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643      2.222473
#>                                                      subject_No.15
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     0.6398326
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643    -3.5834583
#>                                                      subject_No.16
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872    -3.0828362
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643    -0.2547361
#>                                                      subject_No.17
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872      4.565988
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     -1.607503
#>                                                      subject_No.18
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     -4.559211
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     -2.126832
#>                                                      subject_No.19
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     -1.650555
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     -2.628202
#>                                                      subject_No.20
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     -1.446678
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643      1.331196
#>                                                      subject_No.21
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872      2.252430
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643      1.065067
#>                                                      subject_No.22
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872    -0.7077453
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643    -2.8368271
#>                                                      subject_No.23
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     0.0807707
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     0.1465928
#>                                                      subject_No.24
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     -1.801513
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     -3.531221
#>                                                      subject_No.25
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872      2.429377
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643      1.248840
#>                                                      subject_No.26
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872    -0.1593598
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     1.9840856
#>                                                      subject_No.27
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     2.2100434
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     0.6000955
#>                                                      subject_No.28
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872      -1.22412
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643      -1.89099
#>                                                      subject_No.29
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872    0.07799926
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643    0.25674549
#>                                                      subject_No.30
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     -1.766478
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     -1.290206
#>                                                      subject_No.31
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872    -3.1370985
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     0.7140074
#>                                                      subject_No.32
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872      3.251277
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643      1.136505
#>                                                      subject_No.33
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     -3.669775
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643      2.290494
#>                                                      subject_No.34
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872      1.490325
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643      2.988986
#>                                                      subject_No.35
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872   0.188411662
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643  -0.003828624
#>                                                      subject_No.36
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     -2.702655
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643      2.134315
#>                                                      subject_No.37
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872      1.305235
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     -1.949074
#>                                                      subject_No.38
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     0.8252565
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     2.8331756
#>                                                      subject_No.39
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     4.7394221
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643    -0.5380792
#>                                                      subject_No.40
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     0.1082640
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     0.0982518
#>                                                      subject_No.41
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     4.1195389
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643    -0.7289868
#>                                                      subject_No.42
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     0.2144923
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643    -3.1639794
#>                                                      subject_No.43
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872    -0.5894508
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643    -0.8353333
#>                                                      subject_No.44
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872      3.256380
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643      3.352047
#>                                                      subject_No.45
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     0.4212448
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     4.1328608
#>                                                      subject_No.46
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     -2.162430
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     -1.268348
#>                                                      subject_No.47
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     0.9935125
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643    -3.0822171
#>                                                      subject_No.48
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     -3.446872
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     -1.304178
#>                                                      subject_No.49
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872    0.27131545
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643   -0.01378271
#>                                                      subject_No.50
#> dataset_No.1_subcomp.1, PVE: 0.055324, PVE: 0.001872     -2.769463
#> dataset_No.1_subcomp.2, PVE: 0.047130, PVE: 0.001643     -2.832879
#>