Percentage of Variance Explained for separate data set

Compute the PVE (percentage of variance explained) for each data set

Usage

pveSep(dataset, list_score, list_component)

Arguments

dataset

A list of data sets for input

list_score

A list of extracted scores by the corresponding algorithm

list_component

A list of components comptuted by the corresponding algorithm

Value

The list of scores

Examples

dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
comp_num = 2
res_sepPCA = sepPCA(dataset, comp_num)
pveSep(dataset, res_sepPCA$score_list, res_sepPCA$linked_component_list)
#> $dataset_No.1
#>                                                      subject_No.1 subject_No.2
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  0.004394594 -0.001287893
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369 -0.002126189  0.008618308
#>                                                      subject_No.3 subject_No.4
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275 -0.005220780 -0.001187708
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369 -0.004702208 -0.001955929
#>                                                      subject_No.5 subject_No.6
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  0.001770701 -0.003732382
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369 -0.004288846 -0.002200328
#>                                                      subject_No.7 subject_No.8
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275 -0.002640748 0.0004423825
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  0.003775150 0.0012512872
#>                                                      subject_No.9 subject_No.10
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275 -0.001945232  -0.001081409
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369 -0.002384727  -0.005117146
#>                                                      subject_No.11
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  -0.004685800
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369   0.001843405
#>                                                      subject_No.12
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275   0.003471540
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.002404352
#>                                                      subject_No.13
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275   0.002601323
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.002680696
#>                                                      subject_No.14
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  0.0034552414
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  0.0006125964
#>                                                      subject_No.15
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  -0.002328091
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.003258443
#>                                                      subject_No.16
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275 -0.0057288370
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  0.0005670049
#>                                                      subject_No.17
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  -0.004065531
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369   0.007291599
#>                                                      subject_No.18
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  -0.007728402
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.003623109
#>                                                      subject_No.19
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275 -0.0008467369
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  0.0012577910
#>                                                      subject_No.20
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  -0.001193285
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.001478380
#>                                                      subject_No.21
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275   0.004907569
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369   0.003497420
#>                                                      subject_No.22
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275   0.001492583
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369   0.003740335
#>                                                      subject_No.23
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  -0.006132507
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369   0.003534345
#>                                                      subject_No.24
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  -0.003804764
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.004860444
#>                                                      subject_No.25
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  -0.006391925
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.002572812
#>                                                      subject_No.26
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275   0.004211881
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369   0.005472646
#>                                                      subject_No.27
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275 -0.0004802451
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  0.0046016746
#>                                                      subject_No.28
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275   0.003331680
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369   0.003654587
#>                                                      subject_No.29
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  0.0094805893
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  0.0006752269
#>                                                      subject_No.30
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  0.0007494947
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369 -0.0073410158
#>                                                      subject_No.31
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275   0.002859656
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.001309305
#>                                                      subject_No.32
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  0.0036247972
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369 -0.0006690951
#>                                                      subject_No.33
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  -0.002036014
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369   0.004096751
#>                                                      subject_No.34
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  -0.001732418
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.005290089
#>                                                      subject_No.35
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  -0.005044253
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369   0.003537337
#>                                                      subject_No.36
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  -0.003573450
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369   0.006291253
#>                                                      subject_No.37
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275 -0.0008741825
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369 -0.0064538277
#>                                                      subject_No.38
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  0.0004135267
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  0.0072470303
#>                                                      subject_No.39
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275   0.005072625
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.008252264
#>                                                      subject_No.40
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  0.0044103816
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369 -0.0005270218
#>                                                      subject_No.41
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275   0.002327541
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.004777147
#>                                                      subject_No.42
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  -0.001405549
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.005997668
#>                                                      subject_No.43
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275   0.002871648
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369   0.008565636
#>                                                      subject_No.44
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275   0.001559806
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.001497099
#>                                                      subject_No.45
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  -0.003724270
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369   0.002822902
#>                                                      subject_No.46
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275 -0.0003152679
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  0.0038796023
#>                                                      subject_No.47
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275  -0.008338442
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.002725273
#>                                                      subject_No.48
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275   0.004219900
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369   0.004425368
#>                                                      subject_No.49
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275   0.003852261
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.001395996
#>                                                      subject_No.50
#> dataset_No.1_subcomp.1, PVE: 0.054426, PVE: 0.010275   0.016004399
#> dataset_No.1_subcomp.2, PVE: 0.050079, PVE: 0.002369  -0.001369844
#>