Percentage of Variance Explained for separate data set

Compute the PVE (percentage of variance explained) for each data set

pveSep(dataset, list_score, list_component)

Arguments

dataset
A list of data sets for input

list_score
A list of extracted scores by the corresponding algorithm

list_component
A list of components comptuted by the corresponding algorithm

Value

The list of scores

Examples

dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
comp_num = 2
res_sepPCA = sepPCA(dataset, comp_num)
pveSep(dataset, res_sepPCA$score_list, res_sepPCA$linked_component_list)
#> $dataset_No.1
#>                                                      subject_No.1 subject_No.2
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865   -1.3575851     3.297226
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818    0.7153427     1.251888
#>                                                      subject_No.3 subject_No.4
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     1.268366   -0.1677864
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     3.073735   -0.4991476
#>                                                      subject_No.5 subject_No.6
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865    0.6791097    0.8708951
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818   -4.7229399   -2.1906876
#>                                                      subject_No.7 subject_No.8
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865    -3.001422    -3.563259
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818    -1.723859    -1.647770
#>                                                      subject_No.9 subject_No.10
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865    0.9350738     -2.591253
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818   -1.0403555     -3.846145
#>                                                      subject_No.11
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     4.7108776
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818    -0.1318122
#>                                                      subject_No.12
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     0.1541386
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     0.5578099
#>                                                      subject_No.13
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     0.3509357
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818    -0.5607943
#>                                                      subject_No.14
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     0.7008346
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818    -0.7909920
#>                                                      subject_No.15
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     2.7637516
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     0.5408017
#>                                                      subject_No.16
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865      1.185401
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818      1.040041
#>                                                      subject_No.17
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865      2.735432
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818      3.324290
#>                                                      subject_No.18
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865    -0.6827872
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818    -0.4833946
#>                                                      subject_No.19
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865    -0.5576739
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818    -1.9895517
#>                                                      subject_No.20
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     -1.843170
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818      3.740772
#>                                                      subject_No.21
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     4.2490904
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818    -0.6274634
#>                                                      subject_No.22
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865   -2.92205917
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818    0.08244529
#>                                                      subject_No.23
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865      1.099250
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     -2.403155
#>                                                      subject_No.24
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865      3.711698
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     -2.634315
#>                                                      subject_No.25
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     0.1123962
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     2.2220942
#>                                                      subject_No.26
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865    -0.3063353
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818    -0.7197032
#>                                                      subject_No.27
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865      4.374969
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     -4.281391
#>                                                      subject_No.28
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     -1.955966
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818      2.182740
#>                                                      subject_No.29
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     1.6460393
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     0.7242648
#>                                                      subject_No.30
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     0.2959862
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     2.9203636
#>                                                      subject_No.31
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865   -0.95068941
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818    0.05727601
#>                                                      subject_No.32
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     -4.926835
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818      1.191988
#>                                                      subject_No.33
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     -1.012540
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818      2.754229
#>                                                      subject_No.34
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     -3.524852
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818      2.856965
#>                                                      subject_No.35
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     1.2772327
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818    -0.9212174
#>                                                      subject_No.36
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     0.4530733
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     1.5401087
#>                                                      subject_No.37
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865      3.504615
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818      2.035872
#>                                                      subject_No.38
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865    -1.8674963
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818    -0.1800033
#>                                                      subject_No.39
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865    -1.7207393
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818    -0.3084279
#>                                                      subject_No.40
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865    -2.4427592
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     0.8725306
#>                                                      subject_No.41
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865    -2.6044141
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     0.8785756
#>                                                      subject_No.42
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865      3.033886
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818      2.903103
#>                                                      subject_No.43
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     -1.353543
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     -5.133424
#>                                                      subject_No.44
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865      2.847384
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     -1.768128
#>                                                      subject_No.45
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     -2.044562
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     -3.715437
#>                                                      subject_No.46
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     -2.056483
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     -2.825513
#>                                                      subject_No.47
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865      1.297897
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818      3.591339
#>                                                      subject_No.48
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865    -2.2034624
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818    -0.1675603
#>                                                      subject_No.49
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865    -0.9876808
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818     2.7982470
#>                                                      subject_No.50
#> dataset_No.1_subcomp.1, PVE: 0.053400, PVE: 0.001865     -1.213755
#> dataset_No.1_subcomp.2, PVE: 0.051021, PVE: 0.001818      1.692587
#>