Percentage of Variance Explained for separate data set

Compute the PVE (percentage of variance explained) for each data set

Usage

pveSep(dataset, list_score, list_component)

Arguments

dataset

A list of data sets for input

list_score

A list of extracted scores by the corresponding algorithm

list_component

A list of components comptuted by the corresponding algorithm

Value

The list of scores

Examples

dataset = list(matrix(runif(5000, 1, 2), nrow = 100, ncol = 50))
comp_num = 2
res_sepPCA = sepPCA(dataset, comp_num)
pveSep(dataset, res_sepPCA$score_list, res_sepPCA$linked_component_list)
#> $dataset_No.1
#>                                                      subject_No.1 subject_No.2
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -0.003425507 -0.006847713
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  0.002899917  0.003888993
#>                                                      subject_No.3  subject_No.4
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  0.002317228 -6.277743e-05
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  0.001927201  3.618940e-03
#>                                                      subject_No.5 subject_No.6
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -0.004245341 -0.002354755
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997 -0.006464237 -0.001864531
#>                                                       subject_No.7 subject_No.8
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -0.0021464317 -0.001467248
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997 -0.0007982595  0.007300671
#>                                                       subject_No.9
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -0.0006225371
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997 -0.0010631515
#>                                                      subject_No.10
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -9.928557e-05
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997 -1.081285e-03
#>                                                      subject_No.11
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  -0.007021139
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997   0.004676304
#>                                                      subject_No.12
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162   0.004701768
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  -0.002346261
#>                                                      subject_No.13
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  0.0029952769
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  0.0002996064
#>                                                      subject_No.14
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  -0.007327703
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  -0.006171503
#>                                                      subject_No.15
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  -0.002128243
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997   0.003144191
#>                                                      subject_No.16
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  -0.001083186
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  -0.002691274
#>                                                      subject_No.17
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162   0.007894644
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997   0.002862767
#>                                                      subject_No.18
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162   0.009208314
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997   0.001985936
#>                                                      subject_No.19
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  0.0009300753
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997 -0.0014134685
#>                                                      subject_No.20
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -0.0021814921
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997 -0.0009679307
#>                                                      subject_No.21
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  0.0050295474
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997 -0.0009903695
#>                                                      subject_No.22
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162   0.011805382
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997   0.001534154
#>                                                      subject_No.23
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  -0.004078256
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997   0.003490045
#>                                                      subject_No.24
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -0.0010652469
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  0.0006951659
#>                                                      subject_No.25
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162   0.006738935
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997   0.003477057
#>                                                      subject_No.26
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  -0.003207799
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  -0.012156740
#>                                                      subject_No.27
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -0.0002629557
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  0.0032054450
#>                                                      subject_No.28
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -0.0004303825
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997 -0.0023693971
#>                                                      subject_No.29
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -0.0048479680
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  0.0004981029
#>                                                      subject_No.30
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -0.0032668224
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  0.0003719025
#>                                                      subject_No.31
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162   0.003220840
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997   0.002079313
#>                                                      subject_No.32
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  -0.001641865
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  -0.009322078
#>                                                      subject_No.33
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  -0.001417000
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997   0.007723585
#>                                                      subject_No.34
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -0.0018055449
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997 -0.0006172419
#>                                                      subject_No.35
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162   0.001953438
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997   0.008137885
#>                                                      subject_No.36
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  -0.005949581
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997   0.004797167
#>                                                      subject_No.37
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -0.0007134835
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997 -0.0033153560
#>                                                      subject_No.38
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  -0.004485809
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  -0.005112901
#>                                                      subject_No.39
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -0.0008342771
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997 -0.0052834915
#>                                                      subject_No.40
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -0.0004212729
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  0.0027046729
#>                                                      subject_No.41
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  -0.002158810
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997   0.005372234
#>                                                      subject_No.42
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  0.0062290206
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  0.0003739308
#>                                                      subject_No.43
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162   0.001690072
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  -0.006110330
#>                                                      subject_No.44
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -1.577284e-05
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  5.728053e-03
#>                                                      subject_No.45
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162 -0.0047855575
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997 -0.0003236792
#>                                                      subject_No.46
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162   0.004264497
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  -0.007402364
#>                                                      subject_No.47
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  0.0061272296
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997 -0.0004170099
#>                                                      subject_No.48
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162  -0.004071175
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997   0.002840347
#>                                                      subject_No.49
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162   0.002376404
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  -0.002001249
#>                                                      subject_No.50
#> dataset_No.1_subcomp.1, PVE: 0.053160, PVE: 0.011162   0.008990265
#> dataset_No.1_subcomp.2, PVE: 0.051657, PVE: 0.002997  -0.005349482
#>